Created a differnt component for singlePost
