+++ fe/src/index.css
@@ -31,7 +31,7 @@ code {
 
   --success: #1fbf43;
   --error: #e64034;
- 
+
   --shadow-1: -6px 6px 10px 4px rgb(0 0 0/ 0.35);
   --shadow-2: -2px 2px 10px rgb(0 0 0/ 0.35);
   /* Text-size styles */
@@ -63,4 +63,8 @@ code {
   background: var(--primary) !important;
   color: white !important;
   font-weight: 700;
+}
+
+a {
+  text-decoration: none;
 }
