+++ fe/src/pages/EditProfilePage/index.jsx
@@ -1,7 +1,11 @@
 import React, { useEffect, useRef, useState } from "react";
 import { Tab, Tabs, TabList, TabPanel } from "react-tabs";
 import "react-tabs/style/react-tabs.css";
-import { authenticAxiosInstance, getCookie } from "../../utils/axios";
+import {
+  authenticAxiosInstance,
+  getCookie,
+  getCurrentUserId,
+} from "../../utils/axios";
 import {
   checkForInvalidOrExpiredToken,
   decodeJWT,
@@ -12,53 +16,37 @@ import { format, parseISO } from "date-fns";
 import style from "./profile.module.css";
 import CustomButton from "../../components/Button";
 import toast from "react-hot-toast";
-import { json, useNavigate } from "react-router-dom";
+import { json, Link, useNavigate } from "react-router-dom";
 import WriteIcon from "../../components/icons/WriteIcon";
 import Modal from "../../components/Modal";
 import CustomInput from "../../components/Input";
 import { getUsers } from "../../utils/Users";
 import UserBox from "../../components/UserBox";
+import SinglePost from "../../components/SinglePost";
 
-const formatDate = (dateString) => {
-  // Parse the string to a Date object and format it
-  const date = new Date(dateString);
-  return format(date, "yyyy-MM-dd");
-};
 
 const EditProfilePage = () => {
   const [user, setUser] = useState();
   const [followers, setFollowers] = useState(null);
   const [tweets, setTweets] = useState(null);
-
   const [isOpenFollowerModal, setIsOpenFollowerModal] = useState(false);
-
   const navigate = useNavigate();
 
-  // console.log("Request Headers:", authenticAxiosInstance.defaults.headers);
-
   useEffect(() => {
-    const token = getCookie("tokenFromServer");
+    const currentId = getCurrentUserId("tokenFromServer");
 
     try {
-      if (token) {
-        console.log("Token ", token);
-        const data = decodeJWT(token);
-        console.log("Token ", data);
-        console.log(data.sub);
-        console.log(data.csrf);
-        console.log("Token ", data.fresh);
-
+      if (currentId) {
         authenticAxiosInstance
-          .get("/users/" + data.sub)
+          .get("/users/" + currentId)
           .then(function (response) {
             console.log(response);
             if (response.data) {
               if (response.data.status === "failure") {
                 toast.error(response.data.data);
-                // setIncorrectData(response.data.data);
               } else {
                 toast.success("User " + response.data.status);
-                console.log(response.data.data);
+                console.log("Getting user ", response.data.data);
 
                 setUser(response.data.data);
               }
@@ -68,7 +56,6 @@ const EditProfilePage = () => {
           })
           .catch(function (error) {
             console.log("User error ", error);
-
             toast.error("Something Went Wrong.");
           });
 
@@ -107,7 +94,7 @@ const EditProfilePage = () => {
 
         // for tweets
         authenticAxiosInstance
-          .get(`/users/${data.sub}/tweets`)
+          .get(`/users/${currentId}/tweets`)
           .then(function (response) {
             console.log(response);
             if (response.data) {
@@ -185,24 +172,6 @@ const EditProfilePage = () => {
           }
         });
     }, []);
-
-    // const [followesData, setfollowesData] = useState([]);
-
-    // useEffect(() => {
-    //   //TODO
-    //   setfollowesData([]);
-
-    //   const idList = followers.map((item) => item.id);
-
-    //   // Filter users whose id is not in the ids array
-    //   const usersNotInList = users.filter((user) => !idList.includes(user.id));
-
-    //   console.log("followers ", usersNotInList);
-
-    //   setfollowesData(usersNotInList);
-    // }, [users]);
-
-    // console.log(followesData);
 
     return (
       <Tabs>
@@ -296,7 +265,7 @@ const EditProfilePage = () => {
           {user &&
             tweets &&
             tweets.map((item) => {
-              return <SinglePost user={user} tweet={item} />;
+              return <SinglePost user={user} tweet={item} currentUser />;
             })}
         </div>
       </div>
@@ -304,242 +273,5 @@ const EditProfilePage = () => {
   );
 };
 
-const DeleteModal = ({ id, onClose }) => {
-  const navigate = useNavigate();
-  const deleteTweetAPI = () =>
-    authenticAxiosInstance
-      .delete(`/tweets/${id}`)
-      .then(function (response) {
-        console.log(response);
-        if (response.data) {
-          if (response.data.status === "failure") {
-            toast.error(response.data.data);
-          } else {
-            toast.success("Tweets: " + response.data.status);
-            console.log("Tweets", response.data.data);
-            window.location.reload();
-            onClose();
-          }
-        } else {
-          toast.error("Tweets: Something Went Wrong.");
-        }
-      })
-      .catch(function (error) {
-        if (error.response) {
-          console.log(error.response.data);
-
-          let invalid = checkForInvalidOrExpiredToken(error.response.data.msg);
-
-          if (invalid) {
-            navigate("/login");
-            toast.error("You need to login");
-          } else {
-            console.log("Error 0", error.response.data);
-          }
-        }
-      });
-
-  return (
-    <div>
-      <h2>Would you like to delete?</h2>
-      <div>
-        <button
-          onClick={() => {
-            deleteTweetAPI();
-          }}
-        >
-          Yes
-        </button>
-        <button
-          onClick={() => {
-            onClose();
-          }}
-        >
-          No
-        </button>
-      </div>
-    </div>
-  );
-};
-
-const UpdateModal = ({ tweet, onCancel }) => {
-  const [content, setContent] = useState(tweet.content ? tweet.content : "");
-
-  const navigate = useNavigate();
-
-  const updateModalAPI = () =>
-    authenticAxiosInstance
-      .patch(`/tweets/${tweet.id}`, {
-        content,
-      })
-      .then(function (response) {
-        console.log(response);
-        if (response.data) {
-          if (response.data.status === "failure") {
-            toast.error(response.data.data);
-          } else {
-            toast.success("Tweets: " + response.data.status);
-            console.log("Tweets", response.data.data);
-            window.location.reload();
-            onCancel();
-          }
-        } else {
-          toast.error("Tweets: Something Went Wrong.");
-        }
-      })
-      .catch(function (error) {
-        if (error.response) {
-          console.log(error.response.data);
-
-          let invalid = checkForInvalidOrExpiredToken(error.response.data.msg);
-
-          if (invalid) {
-            navigate("/login");
-            toast.error("You need to login");
-          } else {
-            console.log("Error 0", error.response.data);
-          }
-        }
-      });
-
-  return (
-    <div>
-      <CustomInput
-        value={content}
-        hint={"Enter Content"}
-        onChange={(e) => {
-          setContent(e.target.value);
-        }}
-      />
-      <button
-        onClick={() => {
-          updateModalAPI();
-        }}
-      >
-        Update
-      </button>
-      <button
-        onClick={() => {
-          onCancel();
-        }}
-      >
-        Cancel
-      </button>
-    </div>
-  );
-};
-
-export const SinglePost = ({ user, tweet }) => {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const [openDeleteConfirmBox, setOpenDeleteConfirmBox] = useState(false);
-  const [openUpdateModal, setOpenUpdateModal] = useState(false);
-
-  const modalRef = useRef(null); // Ref for modal
-
-  console.log(console.log("Given ", user));
-
-  // Function to handle click outside the modal
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (modalRef.current && !modalRef.current.contains(event.target)) {
-        setIsOpen(false); // Close the modal if clicked outside
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
-
-  return (
-    <>
-      <Modal
-        isOpen={openDeleteConfirmBox}
-        onClose={() => {
-          setOpenDeleteConfirmBox(false);
-        }}
-        children={
-          <DeleteModal
-            id={tweet.id}
-            onClose={() => {
-              setOpenDeleteConfirmBox(false);
-            }}
-          />
-        }
-      />
-
-      <Modal
-        isOpen={openUpdateModal}
-        onClose={() => {
-          setOpenUpdateModal(false);
-        }}
-        children={
-          <UpdateModal
-            tweet={tweet}
-            onCancel={() => {
-              setOpenUpdateModal(false);
-            }}
-          />
-        }
-      />
-
-      <div className={`${style.singlePost}`}>
-        <div className={`${style.singleCircularProfile}`}>
-          {user && user.username && user.username.slice(0, 1).toUpperCase()}
-        </div>
-        <div className={style.postRightContainer}>
-          <div className={style.singlePostHeaderContainer}>
-            <div className={style.singlePostHeader}>
-              <h2>{user.username}</h2>
-              <p>{formatDate(tweet.created_at)}</p>
-            </div>
-            <div className={style.dropdown}>
-              <div
-                onClick={() => {
-                  setIsOpen(true);
-                }}
-              >
-                <WriteIcon />
-              </div>
-              {isOpen && (
-                <div className={style.modal} ref={modalRef}>
-                  <div
-                    style={{
-                      display: "flex",
-                      justifyContent: "center",
-                      alignItems: "center",
-                      cursor: "pointer",
-                    }}
-                    onClick={() => {
-                      setOpenUpdateModal(true);
-                    }}
-                  >
-                    Update <WriteIcon />
-                  </div>
-                  <div
-                    style={{
-                      display: "flex",
-                      justifyContent: "center",
-                      alignItems: "center",
-                      cursor: "pointer",
-                    }}
-                    onClick={() => {
-                      setOpenDeleteConfirmBox(true);
-                    }}
-                  >
-                    Delete <WriteIcon />
-                  </div>
-                </div>
-              )}
-            </div>
-          </div>
-          <p>{tweet.content}</p>
-        </div>
-      </div>
-    </>
-  );
-};
 
 export default EditProfilePage;
